Refactor lightmap part of shader generation by the standard materials
