+++ src/scene/shader-lib/chunks/chunks.js
@@ -69,8 +69,7 @@ import lightFunctionLightPS from './lit/frag/lighting/lightFunctionLight.js';
 import lightFunctionShadowPS from './lit/frag/lighting/lightFunctionShadow.js';
 import lightingPS from './lit/frag/lighting/lighting.js';
 import lightmapAddPS from './lit/frag/lightmapAdd.js';
-import lightmapDirPS from './standard/frag/lightmapDir.js';
-import lightmapSinglePS from './standard/frag/lightmapSingle.js';
+import lightmapPS from './standard/frag/lightmap.js';
 import lightSpecularAnisoGGXPS from './lit/frag/lightSpecularAnisoGGX.js';
 import lightSpecularBlinnPS from './lit/frag/lightSpecularBlinn.js';
 import lightSheenPS from './lit/frag/lightSheen.js';
@@ -279,8 +278,7 @@ const shaderChunks = {
     lightFunctionShadowPS,
     lightingPS,
     lightmapAddPS,
-    lightmapDirPS,
-    lightmapSinglePS,
+    lightmapPS,
     lightSpecularAnisoGGXPS,
     lightSpecularBlinnPS,
     lightSheenPS,
